# contracts/education/openapi.yaml
openapi: 3.0.3
info:
  title: Education Service API
  version: 0.1.0
  description: |
    学员服务 BFF/业务 API 契约（对应 specs/education-service-agent/design.md §5）。
    约定：JSON camelCase；金额为整数分；时间为 ISO8601 UTC；studentId 永远由服务端从会话取得，
    请求体中不出现。写操作接受 Idempotency-Key；状态更新携带 expectedRevision。
servers:
  - url: /api/v1
...
paths:
...
  /conversations/{conversationId}/handoff:
    post:
      tags: [chat]
      operationId: requestHandoff
      summary: 学员请求老师接管：会话 bot → queued
      description: |
        幂等：会话已经在排队或接管中，返回那条进行中的记录（不新建、不覆盖原因）。会话已关闭返回 409 INVALID_STATE。
        Agent 认为该转人工时不直接调用这个接口（/conversations/* 对 Agent 通道关闭）：它通过事件流告诉 BFF，由 BFF 落库。
      parameters: [{$ref: '#/components/parameters/ConversationId'}]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string, maxLength: 500}
      responses:
        '200':
          description: 队列状态
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Handoff'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /teacher/handoffs:
    get:
      tags: [teacher]
      operationId: listTeacherHandoffs
      summary: 进行中的接管队列（待接管 + 已接管），供老师工作台展示
      description: |
        只返回排队中（queued）和已接管（claimed）的记录，按发起时间排序；已结束（released）的不在这里。
        每条都带交接摘要（Agent 转人工时写的；学员直接点按钮转人工的没有摘要）。
      responses:
        '200':
          description: 队列
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/Handoff'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /teacher/handoffs/{handoffId}/claim:
    post:
      tags: [teacher]
      operationId: claimHandoff
      summary: 老师接管排队中的记录：queued → claimed，会话 queued → human
      description: |
        接管不是幂等的：记录已被接管（含同一位老师再点一次）返回 409 INVALID_STATE；expectedRevision 过期返回 409 REVISION_CONFLICT。
        两位老师同时接管同一条，恰好一个成功。
      parameters: [{$ref: '#/components/parameters/HandoffId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExpectedRevisionRequest'}
      responses:
        '200':
          description: 已接管
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Handoff'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /teacher/handoffs/{handoffId}/release:
    post:
      tags: [teacher]
      operationId: releaseHandoff
      summary: 接管人结束接管：claimed → released，会话 human → bot
      description: |
        只有接管这条记录的老师能结束（别的老师 403）；记录不在 claimed 状态返回 409 INVALID_STATE；expectedRevision 过期返回 409 REVISION_CONFLICT。
      parameters: [{$ref: '#/components/parameters/HandoffId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExpectedRevisionRequest'}
      responses:
        '200':
          description: 已结束接管，会话回到 bot
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Handoff'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /teacher/conversations/{conversationId}/messages:
    get:
      tags: [teacher]
      operationId: listConversationMessagesForTeacher
      summary: 已接管会话的消息与交接摘要，供老师工作台展示；也用于接管期间轮询学员消息
      description: |
        只有当前接管这个会话的老师能看（最小暴露：不能翻看别人接管过、或还在排队没人接管的会话）。
        不在你的接管中（排队中、被别人接管、或早已回到 bot）返回 409 INVALID_STATE；被另一位老师接管返回 403。
      parameters:
        - {$ref: '#/components/parameters/ConversationId'}
        - name: after
          in: query
          required: false
          description: 同 GET /conversations/{conversationId}/messages 的 after：不传是最近 200 条快照，传某条消息 id 是它之后的增量。
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: 消息与交接摘要
          content:
            application/json:
              schema:
                type: object
                required: [items, handoff]
                properties:
                  items:
                    description: 不传 after 时是最近 200 条快照；传 after 时是它之后的新消息（可能为空数组）。
                    type: array
                    items: {$ref: '#/components/schemas/Message'}
                  handoff: {$ref: '#/components/schemas/Handoff'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
    post:
      tags: [teacher]
      operationId: teacherSendMessage
      summary: 接管中的老师发消息
      description: |
        只有当前接管这个会话的老师能发：会话不在人工接管中返回 409 INVALID_STATE，被另一位老师接管返回 403。
        同一个 clientMessageId 重试返回 200 和原消息（不再写第二条）；clientMessageId 已被另一种消息占用返回 422。
      parameters: [{$ref: '#/components/parameters/ConversationId'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientMessageId, text]
              properties:
                clientMessageId: {type: string, minLength: 1, maxLength: 100}
                text: {type: string, minLength: 1, maxLength: 4000}
      responses:
        '200':
          description: 重复的 clientMessageId，返回原消息
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Message'}
        '201':
          description: 已发送
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Message'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
  parameters:
...
    ConversationId: {name: conversationId, in: path, required: true, schema: {type: string, format: uuid}}
    HandoffId: {name: handoffId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
...
    Forbidden:
      description: 角色无权限
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: 不存在，或他人私人资源（不区分）
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Invalid:
      description: 参数错误
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Conflict:
      description: 状态或版本冲突
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, INVALID_STATE, REVISION_CONFLICT,
                     STALE_CONFIRMATION, IDEMPOTENCY_KEY_REUSED, RUN_IN_PROGRESS, HANDOFF_ACTIVE,
                     BUDGET_EXCEEDED, DEPENDENCY_UNAVAILABLE, INTERNAL]
            message: {type: string}
            requestId: {type: string}
            details: {type: object, additionalProperties: true}
...
    ExpectedRevisionRequest:
      type: object
      required: [expectedRevision]
      properties:
        expectedRevision: {type: integer, minimum: 0}
...
    Message:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id: {type: string}
        role: {type: string, enum: [user, assistant, teacher, system]}
        content: {type: string}
        runId: {type: string, nullable: true}
        createdAt: {type: string, format: date-time}
    Handoff:
      type: object
      required: [id, conversationId, status, revision]
      properties:
        id: {type: string, format: uuid}
        conversationId: {type: string, format: uuid}
        status: {type: string, enum: [queued, claimed, released]}
        teacherId: {type: string, format: uuid, nullable: true}
        summary:
          type: string
          nullable: true
          description: |
            交接摘要，写给老师看的内部说明：由 Agent 用确定性代码从已核验的工具结果和学员原话拼出，不含猜测。
            **只在老师接口（claim / release）的响应里返回**；学员侧接口（POST handoff、GET messages）不含这个字段。
            学员直接点"转人工"、没有 Agent 参与时为 null。
        reason: {type: string, nullable: true}
        revision: {type: integer}
security:
  - sessionCookie: []